fix: use the updated initialization function in delegator
Here I created a GovernorBravoDelegateInterface contract that is
imported in the Delegator contract. The initialization selector is
determined based on the interface, so when and if the implementation
initialization changes again, we will remember to update the
delegator contract as well (not on chain but at least in the repo).